DEV-5832: ad correct_bond_price_quotes()
- add test for correct_bond_price_quotes()
- move command line app utility tests to unit test folder
