+++ tests/unit/apps/test_data/mapping_invalid.json
